Added patches to allow a cross-compiler to be compiled.

+++ floppy.h
@@ -4,6 +4,6 @@
 
 unsigned int sector_size;	//the size of the sector in bytes
 
-unsigned int sector_buffer;
+unsigned long sector_buffer;
 	//this is used so that a specialized memory allocater will not be required
 	//but it is required that this buffer be allocated early so that it can claim a spot in lower memory
